gen: set omitempty in json tags for null types
On 21 July 2023, a updateRepositoryRuleset mutation setting a
RuleParametersInput was observed to fail with the error
"Only one rule parameter type can be specified.", even though exactly
on field of RuleParametersInput was non-null. GitHub seems to require
that exacty one field is _present_, not that exactly one field is
non-null.

The generated Go struct for RuleParametersInput is not very useful,
since it does not include omitempty in JSON tags and as such _never_
marshals into JSON accepted by GitHub.

This commit fixes this.